Initial changes for phase II
Initial phase II changes include redefining endpoint to have an endpoint
id, apploications (with id) for an endpoint, role as sender, receiver,
or both of an application. Network flows as a collection of unicast
flows and multicast flows. APIs for unicast and multicast network flows.
APIs include GET, POST, PATCH, and DELETE.